cleanups, removed examples
